fix(map): move activity exports to src/ and remove hardcoded JWT
Move 8 source files from supabase/functions/_shared/activity/ into
src/activity/, replacing originals with re-exports so Edge Functions
continue to resolve their imports. Update package.json exports to
point to src/.

Add capture() method to supabase-cli.js and use it in activity start()
to dynamically read the service role key from `supabase status`, removing
the hardcoded demo JWT from source.

Part 02 of spec 410.
