+++ products/map/src/commands/activity.js
@@ -20,14 +20,16 @@ const supabaseCli = createSupabaseCli();
 
 export async function start() {
   await supabaseCli.run(["start"]);
+  const json = await supabaseCli.capture(["status", "--output", "json"]);
+  const status = JSON.parse(json);
   process.stdout.write("\n");
   process.stdout.write(
     formatSubheader("Export these variables to use the activity layer:") +
       "\n\n",
   );
-  process.stdout.write("  export MAP_SUPABASE_URL=http://127.0.0.1:54321\n");
+  process.stdout.write(`  export MAP_SUPABASE_URL=${status.API_URL}\n`);
   process.stdout.write(
-    "  export MAP_SUPABASE_SERVICE_ROLE_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZS1kZW1vIiwicm9sZSI6InNlcnZpY2Vfcm9sZSIsImV4cCI6MTk4MzgxMjk5Nn0.EGIM96RAZx35lJzdJsyH-qQwv8Hdp7fsn3W0YpN81IU\n\n",
+    `  export MAP_SUPABASE_SERVICE_ROLE_KEY=${status.SERVICE_ROLE_KEY}\n\n`,
   );
   return 0;
 }
