+++ products/map/src/lib/supabase-cli.js
@@ -96,5 +96,31 @@ export function createSupabaseCli({
     });
   }
 
-  return { run, resolve };
+  async function capture(args) {
+    const desc = await resolve();
+    if (!desc) {
+      throw new Error(
+        "Could not find the `supabase` CLI. Install it via Homebrew " +
+          "(`brew install supabase/tap/supabase`) or npm " +
+          "(`npm install supabase` in this project, or `npm install -g supabase`), " +
+          `then retry. See ${SUPABASE_INSTALL_URL}.`,
+      );
+    }
+
+    return new Promise((res, rej) => {
+      const chunks = [];
+      const child = spawnFn(desc.cmd, [...desc.prefix, ...args], {
+        cwd,
+        stdio: ["inherit", "pipe", "inherit"],
+      });
+      child.stdout.on("data", (chunk) => chunks.push(chunk));
+      child.on("error", rej);
+      child.on("exit", (code) => {
+        if (code === 0) res(Buffer.concat(chunks).toString());
+        else rej(new Error(`supabase ${args.join(" ")} exited ${code}`));
+      });
+    });
+  }
+
+  return { run, capture, resolve };
 }
